refactor: organizes Stack exports in their respective folder
