+++ src/components/ui/stack/index.ts
@@ -1,0 +1,2 @@
+export { default as Stack } from './Stack.vue';
+export type { StackProps } from './types';
