+++ src/index.ts
@@ -92,13 +92,13 @@ export * from './components/ui/skeleton';
 export * from './components/ui/skeleton-list';
 export * from './components/ui/skeleton-table';
 export * from './components/ui/spinner';
+export * from './components/ui/stack';
 
 // Types
 export * from './types';
 
 export { Toast as $toast } from './components/ui/toast';
 export { default as Toast } from './components/ui/toast/Toast.vue';
-export { default as Stack } from './components/ui/stack/Stack.vue';
 export { default as Badge } from './components/ui/badge/Badge.vue';
 export { default as Tag } from './components/ui/tag/Tag.vue';
 export { default as TagList } from './components/ui/tag/TagList.vue';
